// backend/src/modules/seniors/dtos/update-senior.dto.ts
export class UpdateSeniorDto {
    seniorId?: number;
    firstName?: string;
    middleName?: string;
    lastName?: string;
    suffix?: string;
    age?: number;
    birthMonth?: string;
    birthDay?: number;
    birthYear?: number;
    districtName?: string;
    zone?: string;
    gender?: string;
    email?: string;
    benefits?: string;
  }
  